feat: move LargeSoftCache filter extensions to commons
